+++ src/Pagination/PaginationManipulator.php
@@ -8,8 +8,6 @@ use GraphQL\Language\AST\NonNullTypeNode;
 use GraphQL\Language\AST\ObjectTypeDefinitionNode;
 use GraphQL\Language\AST\TypeNode;
 use GraphQL\Language\Parser;
-use Illuminate\Container\Container;
-use Illuminate\Contracts\Config\Repository as ConfigRepository;
 use Nuwave\Lighthouse\Exceptions\DefinitionException;
 use Nuwave\Lighthouse\Schema\AST\ASTHelper;
 use Nuwave\Lighthouse\Schema\AST\DocumentAST;
@@ -250,20 +248,11 @@ GRAPHQL
         return $description . $definition;
     }
 
-    /**
-     * @return \GraphQL\Language\AST\NamedTypeNode|\GraphQL\Language\AST\NonNullTypeNode
-     */
-    protected function paginationResultType(string $typeName): TypeNode
+    protected function paginationResultType(string $typeName): NonNullTypeNode
     {
-        $config = Container::getInstance()->make(ConfigRepository::class);
-        assert($config instanceof ConfigRepository);
-        $nonNull = $config->get('lighthouse.non_null_pagination_results')
-            ? '!'
-            : '';
-
-        $typeNode = Parser::typeReference(/** @lang GraphQL */ "{$typeName}{$nonNull}");
+        $typeNode = Parser::typeReference(/** @lang GraphQL */ "{$typeName}!");
         assert(
-            $typeNode instanceof NamedTypeNode || $typeNode instanceof NonNullTypeNode,
+            $typeNode instanceof NonNullTypeNode,
             'We do not wrap the typename in [], so this will never be a ListOfTypeNode.'
         );